LIBITD-1088, Updated version.rb version to "1.0.1"
Changed VERSION value in the lib/umd_lib_style/version.rb
file to "1.0.1" so that it matches the Git tag.

Updated the README.md with a reminder to update the
the lib/umd_lib_style/version.rb file when creating new
versions, and additional information for checking out
a particular Git tagged version in projects.

https://issues.umd.edu/browse/LIBITD-1088