LIBITD-1112 Implement "sticky footer" from LIBWEB-4068 in umd_lib_style
Updated umd_lib_style to use "sticky footer" from LIBWEB-4068.

Note: Due to the way the page layout is in umd_lib_style, the
"site" class was placed on the "viewport" div (not the "body" div), and
the "site-content" class placed with the "content" class, not
the div with the "container" class.

Note: The "content" class is used by some Rails applications (such
as Student Applications), so it cannot just be deleted.

Removed old footer sizing CSS and "content" definition from umd_lib.scss,
as it is no longer needed.

Updated documentation.

https://issues.umd.edu/browse/LIBITD-1112