LIBITD-1794. Replaced "HostEnvironmentBanner"
Replaced the "HostEnvironmentBanner" class with a
"DevelopmentEnvironmentBanner" class.

The "HostEnvironmentBanner" class is no longer relevant, because with
the move to Kubernetes, we no longer use the "dev" or "stage" hostnames,
and in a Docker container the host will likely always be "localhost".

Replaced the "HostEnvironmentBanner" class with a
"DevelopmentEnvironmentBanner" class, so that a "Local Environment"
banner would display by default in the local development environment
without any additional configuration.

For test/qa/prod servers, the expectation is that environment variables
will be set to specify the environment banner.

Added unit tests for environment banner.

https://issues.umd.edu/browse/LIBITD-1794