LIBITD-2104. Updated umd_lib_style to use Rails 6.1 and webpacker
Updated to use Rails 6.1 and webpacker, and added "package.json" file
with dependencies.

The major improvement over 2.0.0, is that the
"app/views/layouts/_umd_lib.html.erb" file now uses the
webpacker-compatible "javascript_pack_tag" tag, so that Rails 6
application no longer need to create their own version of the file.

This implementation is likely not quite optimal, because it still
requires the JavaScript dependencies to be added by the Rails
application. However, the fix appears to be to bundle the JavaScript
into a NPM module, and distribute it, which seems to be a little too
involved for this use case, especially as everything will likely
change again in Rails 7.

Replaced the "test/dummy" files with their Rails 6.1 equivalents,
so that the tests would pass.

https://issues.umd.edu/browse/LIBITD-2104